Removed rxJava from ethereum and ethereum-rpc
