+++ ethereum-rpc/build.gradle
@@ -15,9 +15,6 @@ dependencies {
     // Kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlin"
 
-    // Rx
-    implementation "io.reactivex.rxjava2:rxjava:$versions.rxjava"
-
     // Moshi
     implementation "com.squareup.moshi:moshi:$versions.moshi"
     kapt "com.squareup.moshi:moshi-kotlin-codegen:$versions.moshi"
