+++ ethereum/build.gradle
@@ -13,9 +13,6 @@ dependencies {
     // Kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlin"
 
-    // Rx
-    implementation "io.reactivex.rxjava2:rxjava:$versions.rxjava"
-
     // Bivrost
     implementation "com.github.gnosis.bivrost-kotlin:bivrost-solidity-types:$versions.bivrost"
