+++ ethereum/src/main/java/pm/gnosis/ethereum/EthereumRepository.kt
@@ -1,6 +1,5 @@
 package pm.gnosis.ethereum
 
-import io.reactivex.Observable
 import pm.gnosis.ethereum.models.EthereumBlock
 import pm.gnosis.ethereum.models.TransactionData
 import pm.gnosis.ethereum.models.TransactionParameters
@@ -11,30 +10,6 @@ import pm.gnosis.models.Wei
 import java.math.BigInteger
 
 interface EthereumRepository {
-
-    fun <R : BulkRequest> request(bulk: R): Observable<R>
-
-    fun <R : EthRequest<*>> request(request: R): Observable<R>
-
-    fun getBalance(address: Solidity.Address): Observable<Wei>
-
-    fun sendRawTransaction(signedTransactionData: String): Observable<String>
-
-    fun getTransactionReceipt(transactionHash: String): Observable<TransactionReceipt>
-
-    fun getTransactionByHash(transactionHash: String): Observable<TransactionData>
-
-    fun getBlockByHash(blockHash: String): Observable<EthereumBlock>
-
-    fun getTransactionParameters(
-        from: Solidity.Address,
-        to: Solidity.Address,
-        value: Wei? = null,
-        data: String? = null
-    ): Observable<TransactionParameters>
-}
-
-interface CoEthereumRepository {
 
     suspend fun <R : BulkRequest> request(bulk: R): R
